refactor: moving chrome events to the corresponding package
